+++ test/oi/flowgraph_test.exs
@@ -1,0 +1,4 @@
+defmodule Oi.FlowgraphDocTest do
+  use ExUnit.Case, async: true
+  doctest Oi.Flowgraph
+end